requirements.txt
fastapi
uvicorn[standard]
requests
google-auth
psycopg2-binary
pydantic
